+++ src/plone/restapi/pas/plugin.py
@@ -119,6 +119,9 @@ class JWTAuthenticationPlugin(BasePlugin):
 
         userid = payload['sub']
 
+        if isinstance(userid, unicode):
+            userid = userid.encode('utf8')
+
         if self.store_tokens:
             if userid not in self._tokens:
                 return None