+++ CHANGES.rst
@@ -31,6 +31,10 @@ Changelog
 - Fixed NamedFieldDeserializer to validate required fields.
   [jaroel]
 
+- Allow client to ask for the full representation of an object after creation
+  by setting the 'Prefer' header on a PATCH request.
+  [Gagaro]
+
 
 1.0a25 (2017-11-23)
 -------------------