+++ CHANGES.rst
@@ -36,6 +36,13 @@ Changelog
   [Gagaro]
 
 
+Bug Fixes
+
+- Prevent a fatal error when we get @workflow
+  without permission to get review_history worfklow variable.
+  [thomasdesvenain]
+
+
 1.0a25 (2017-11-23)
 -------------------