+++ CHANGES.rst
@@ -35,9 +35,13 @@ Changelog
   by setting the 'Prefer' header on a PATCH request.
   [Gagaro]
 
+- Rename 'url' attribute on navigation / breadcrumb to '@id'.
+  [timo]
+
 - Support deserialization of a relationChoice field using the contents of the
   serialization (enhanced by the serializer) output.
   [sneridagh]
+
 
 
 Bug Fixes